+++ sdk/ml/azure-ai-ml/azure/ai/ml/entities/_compute/compute_instance.py
@@ -138,6 +138,11 @@ class ComputeInstance(Compute):
     :param idle_time_before_shutdown_minutes: Stops compute instance after a user defined period of
         inactivity in minutes. Minimum is 15 min, maximum is 3 days.
     :type idle_time_before_shutdown_minutes: Optional[int], optional
+    :param enable_node_public_ip: Enable or disable node public IP address provisioning. Possible values are:
+        True - Indicates that the compute nodes will have public IPs provisioned.
+        False - Indicates that the compute nodes will have a private endpoint and no public IPs.
+        Default Value: True.
+    :type enable_node_public_ip: Optional[bool], optional
     :param setup_scripts: Details of customized scripts to execute for setting up the cluster.
     :type setup_scripts: Optional[SetupScripts], optional
     """
@@ -157,6 +162,7 @@ class ComputeInstance(Compute):
         idle_time_before_shutdown: Optional[str] = None,
         idle_time_before_shutdown_minutes: Optional[int] = None,
         setup_scripts: Optional[SetupScripts] = None,
+        enable_node_public_ip: Optional[bool] = True,
         **kwargs,
     ):
         kwargs[TYPE] = ComputeType.COMPUTEINSTANCE
@@ -181,6 +187,7 @@ class ComputeInstance(Compute):
         self.idle_time_before_shutdown = idle_time_before_shutdown
         self.idle_time_before_shutdown_minutes = idle_time_before_shutdown_minutes
         self.setup_scripts = setup_scripts
+        self.enable_node_public_ip = enable_node_public_ip
         self.subnet = None
 
     @property
@@ -226,7 +233,9 @@ class ComputeInstance(Compute):
             subnet_resource = ResourceId(id=self.subnet)
         else:
             subnet_resource = None
-        if self.ssh_public_access_enabled and not (self.ssh_settings and self.ssh_settings.ssh_key_value):
+        if self.ssh_public_access_enabled and not (
+            self.ssh_settings and self.ssh_settings.ssh_key_value
+        ):
             msg = "ssh_key_value is required when ssh_public_access_enabled = True."
             raise ValidationException(
                 message=msg,
@@ -240,7 +249,9 @@ class ComputeInstance(Compute):
                 admin_public_key=self.ssh_settings.ssh_key_value,
             )
             if self.ssh_public_access_enabled is not None:
-                ssh_settings.ssh_public_access = "Enabled" if self.ssh_public_access_enabled else "Disabled"
+                ssh_settings.ssh_public_access = (
+                    "Enabled" if self.ssh_public_access_enabled else "Disabled"
+                )
             else:
                 ssh_settings.ssh_public_access = "NotSpecified"
         personal_compute_instance_settings = None
@@ -269,6 +280,13 @@ class ComputeInstance(Compute):
             ssh_settings=ssh_settings,
             personal_compute_instance_settings=personal_compute_instance_settings,
             idle_time_before_shutdown=idle_time_before_shutdown,
+            enable_node_public_ip=self.enable_node_public_ip,
+        )
+        compute_instance_prop.schedules = (
+            self.schedules._to_rest_object() if self.schedules else None
+        )
+        compute_instance_prop.setup_scripts = (
+            self.setup_scripts._to_rest_object() if self.setup_scripts else None
         )
         compute_instance_prop.schedules = (
             self.schedules._to_rest_object() if self.schedules else None
@@ -284,7 +302,9 @@ class ComputeInstance(Compute):
         return ComputeResource(
             location=self.location,
             properties=compute_instance,
-            identity=(self.identity._to_compute_rest_object() if self.identity else None),
+            identity=(
+                self.identity._to_compute_rest_object() if self.identity else None
+            ),
         )
 
     def _to_dict(self) -> Dict:
@@ -331,10 +351,12 @@ class ComputeInstance(Compute):
             network_settings = NetworkSettings(
                 subnet=prop.properties.subnet.id if prop.properties.subnet else None,
                 public_ip_address=prop.properties.connectivity_endpoints.public_ip_address
-                if prop.properties.connectivity_endpoints and prop.properties.connectivity_endpoints.public_ip_address
+                if prop.properties.connectivity_endpoints
+                and prop.properties.connectivity_endpoints.public_ip_address
                 else None,
                 private_ip_address=prop.properties.connectivity_endpoints.private_ip_address
-                if prop.properties.connectivity_endpoints and prop.properties.connectivity_endpoints.private_ip_address
+                if prop.properties.connectivity_endpoints
+                and prop.properties.connectivity_endpoints.private_ip_address
                 else None,
             )
         os_image_metadata = None
@@ -344,8 +366,12 @@ class ComputeInstance(Compute):
                 is_latest_os_image_version=metadata.is_latest_os_image_version
                 if metadata.is_latest_os_image_version is not None
                 else None,
-                current_image_version=metadata.current_image_version if metadata.current_image_version else None,
-                latest_image_version=metadata.latest_image_version if metadata.latest_image_version else None,
+                current_image_version=metadata.current_image_version
+                if metadata.current_image_version
+                else None,
+                latest_image_version=metadata.latest_image_version
+                if metadata.latest_image_version
+                else None,
             )
 
         idle_time_before_shutdown = None
@@ -357,7 +383,9 @@ class ComputeInstance(Compute):
                 pattern=idle_time_before_shutdown_pattern,
                 string=idle_time_before_shutdown,
             )
-            idle_time_before_shutdown_minutes = int(idle_time_match[1]) if idle_time_match else None
+            idle_time_before_shutdown_minutes = (
+                int(idle_time_match[1]) if idle_time_match else None
+            )
 
         response = ComputeInstance(
             name=rest_obj.name,
@@ -381,19 +409,32 @@ class ComputeInstance(Compute):
             create_on_behalf_of=create_on_behalf_of,
             network_settings=network_settings,
             ssh_settings=ssh_settings,
-            ssh_public_access_enabled=_ssh_public_access_to_bool(prop.properties.ssh_settings.ssh_public_access)
-            if (prop.properties and prop.properties.ssh_settings and prop.properties.ssh_settings.ssh_public_access)
+            ssh_public_access_enabled=_ssh_public_access_to_bool(
+                prop.properties.ssh_settings.ssh_public_access
+            )
+            if (
+                prop.properties
+                and prop.properties.ssh_settings
+                and prop.properties.ssh_settings.ssh_public_access
+            )
             else None,
             schedules=ComputeSchedules._from_rest_object(prop.properties.schedules)
-            if prop.properties and prop.properties.schedules and prop.properties.schedules.compute_start_stop
+            if prop.properties
+            and prop.properties.schedules
+            and prop.properties.schedules.compute_start_stop
             else None,
-            identity=IdentityConfiguration._from_compute_rest_object(rest_obj.identity) if rest_obj.identity else None,
+            identity=IdentityConfiguration._from_compute_rest_object(rest_obj.identity)
+            if rest_obj.identity
+            else None,
             setup_scripts=SetupScripts._from_rest_object(prop.properties.setup_scripts)
             if prop.properties and prop.properties.setup_scripts
             else None,
             idle_time_before_shutdown=idle_time_before_shutdown,
             idle_time_before_shutdown_minutes=idle_time_before_shutdown_minutes,
             os_image_metadata=os_image_metadata,
+            enable_node_public_ip=prop.properties.enable_node_public_ip
+            if (prop.properties and prop.properties.enable_node_public_ip)
+            else True,
         )
         return response